fix: rename pipe package->component package
