+++ src/main/java/game/component/util/PipeConstant.java
@@ -1,4 +1,4 @@
-package game.pipe.util;
+package game.component.util;
 
 public final class PipeConstant {
