+++ src/main/java/game/component/util/PipeStatus.java
@@ -1,4 +1,4 @@
-package game.pipe.util;
+package game.component.util;
 
 public enum PipeStatus {
     Left(PipeConstant.LEFT_DRAW_CONSTANT), Right(PipeConstant.RIGHT_DRAW_CONSTANT), Nothing(PipeConstant.NOTHING);
